+++ tests/emailpassword/test_emaildelivery.py
@@ -14,6 +14,8 @@
 
 import json
 from typing import Any, Dict
+from pytest import fixture, mark, raises
+from supertokens_python.exceptions import GeneralError
 
 import httpx
 import respx
@@ -57,6 +59,13 @@ from tests.utils import (
     setup_st,
     sign_up_request,
     start_st,
+)
+from supertokens_python.recipe.emailpassword.asyncio import (
+    send_reset_password_email,
+)
+from supertokens_python.recipe.emailpassword.interfaces import (
+    SendResetPasswordEmailOkResult,
+    SendResetPasswordEmailUnknownUserIdError,
 )
 
 respx_mock = respx.MockRouter
@@ -381,7 +390,7 @@ async def test_reset_password_smtp_service(driver_config_client: TestClient):
 
     def smtp_service_override(oi: SMTPServiceInterface[EmailTemplateVars]):
         async def send_raw_email_override(
-            content: EmailContent, _user_context: Dict[str, Any]
+            content: EmailContent, user_context: Dict[str, Any]
         ):
             nonlocal send_raw_email_called, email
             send_raw_email_called = True
@@ -393,7 +402,7 @@ async def test_reset_password_smtp_service(driver_config_client: TestClient):
             # Note that we aren't calling oi.send_raw_email. So Transporter won't be used.
 
         async def get_content_override(
-            template_vars: EmailTemplateVars, _user_context: Dict[str, Any]
+            template_vars: EmailTemplateVars, user_context: Dict[str, Any]
         ) -> EmailContent:
             nonlocal get_content_called, password_reset_url
             get_content_called = True
@@ -483,7 +492,7 @@ async def test_reset_password_for_non_existent_user(driver_config_client: TestCl
 
     def smtp_service_override(oi: SMTPServiceInterface[EmailTemplateVars]):
         async def send_raw_email_override(
-            content: EmailContent, _user_context: Dict[str, Any]
+            content: EmailContent, user_context: Dict[str, Any]
         ):
             nonlocal send_raw_email_called, email
             send_raw_email_called = True
@@ -495,7 +504,7 @@ async def test_reset_password_for_non_existent_user(driver_config_client: TestCl
             # Note that we aren't calling oi.send_raw_email. So Transporter won't be used.
 
         async def get_content_override(
-            template_vars: EmailTemplateVars, _user_context: Dict[str, Any]
+            template_vars: EmailTemplateVars, user_context: Dict[str, Any]
         ) -> EmailContent:
             nonlocal get_content_called, password_reset_url
             get_content_called = True
@@ -862,6 +871,54 @@ async def test_email_verification_custom_override(driver_config_client: TestClie
 
 
 @mark.asyncio
+async def test_reset_password_backward_compatibility_non_existent_user(
+    driver_config_client: TestClient,
+):
+    "Reset password: test backward compatibility shouldn't send email (non-existent user)"
+    email = ""
+    password_reset_url = ""
+
+    class CustomEmailDeliveryService(
+        emailpassword.EmailDeliveryInterface[emailpassword.EmailTemplateVars]
+    ):
+        async def send_email(
+            self,
+            template_vars: emailpassword.EmailTemplateVars,
+            user_context: Dict[str, Any],
+        ):
+            nonlocal email, password_reset_url
+            email = template_vars.user.email
+            password_reset_url = template_vars.password_reset_link
+
+    init(
+        supertokens_config=SupertokensConfig("http://localhost:3567"),
+        app_info=InputAppInfo(
+            app_name="ST",
+            api_domain="http://api.supertokens.io",
+            website_domain="http://supertokens.io",
+            api_base_path="/auth",
+        ),
+        framework="fastapi",
+        recipe_list=[
+            emailpassword.init(
+                email_delivery=emailpassword.EmailDeliveryConfig(
+                    CustomEmailDeliveryService()
+                )
+            ),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
+        ],
+    )
+    start_st()
+
+    res = reset_password_request(driver_config_client, "test@example.com")
+
+    assert res.status_code == 200
+
+    assert email == ""
+    assert password_reset_url == ""
+
+
+@mark.asyncio
 async def test_email_verification_smtp_service(driver_config_client: TestClient):
     "Email verification: test smtp service"
     email = ""
@@ -874,7 +931,7 @@ async def test_email_verification_smtp_service(driver_config_client: TestClient)
 
     def smtp_service_override(oi: SMTPServiceInterface[VerificationEmailTemplateVars]):
         async def send_raw_email_override(
-            content: EmailContent, _user_context: Dict[str, Any]
+            content: EmailContent, user_context: Dict[str, Any]
         ):
             nonlocal send_raw_email_called, email
             send_raw_email_called = True
@@ -886,7 +943,7 @@ async def test_email_verification_smtp_service(driver_config_client: TestClient)
             # Note that we aren't calling oi.send_raw_email. So Transporter won't be used.
 
         async def get_content_override(
-            template_vars: VerificationEmailTemplateVars, _user_context: Dict[str, Any]
+            template_vars: VerificationEmailTemplateVars, user_context: Dict[str, Any]
         ) -> EmailContent:
             nonlocal get_content_called, email_verify_url
             get_content_called = True
@@ -978,3 +1035,97 @@ async def test_email_verification_smtp_service(driver_config_client: TestClient)
     assert email == "test@example.com"
     assert all([outer_override_called, get_content_called, send_raw_email_called])
     assert email_verify_url
+
+
+@mark.asyncio
+async def test_send_reset_password_email(
+    driver_config_client: TestClient,
+):
+
+    tenant_info, token_info, rid_info, reset_url = "", "", "", ""
+
+    class CustomEmailDeliveryService(
+        emailpassword.EmailDeliveryInterface[PasswordResetEmailTemplateVars]
+    ):
+        async def send_email(
+            self,
+            template_vars: PasswordResetEmailTemplateVars,
+            user_context: Dict[str, Any],
+        ):
+            nonlocal reset_url, token_info, rid_info, tenant_info
+            password_reset_url = template_vars.password_reset_link
+            reset_url = password_reset_url.split("?")[0]
+            token_info = password_reset_url.split("?")[1].split("&")[0]
+            rid_info = password_reset_url.split("?")[1].split("&")[1]
+            tenant_info = password_reset_url.split("?")[1].split("&")[2]
+
+    init(
+        supertokens_config=SupertokensConfig("http://localhost:3567"),
+        app_info=InputAppInfo(
+            app_name="SuperTokens Demo",
+            api_domain="http://api.supertokens.io",
+            website_domain="http://supertokens.io",
+            api_base_path="/auth",
+        ),
+        framework="fastapi",
+        recipe_list=[
+            emailpassword.init(
+                email_delivery=emailpassword.EmailDeliveryConfig(
+                    CustomEmailDeliveryService()
+                )
+            ),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
+        ],
+    )
+    start_st()
+
+    response_1 = sign_up_request(
+        driver_config_client, "random@gmail.com", "validpass123"
+    )
+    assert response_1.status_code == 200
+    dict_response = json.loads(response_1.text)
+    user_info = dict_response["user"]
+    assert dict_response["status"] == "OK"
+    resp = await send_reset_password_email("public", user_info["id"])
+    assert isinstance(resp, SendResetPasswordEmailOkResult)
+
+    assert reset_url == "http://supertokens.io/auth/reset-password"
+    assert token_info is not None and "token=" in token_info
+    assert rid_info is not None and "rid=emailpassword" in rid_info
+    assert tenant_info is not None and "tenantId=public" in tenant_info
+
+
+@mark.asyncio
+async def test_send_reset_password_email_invalid_input(
+    driver_config_client: TestClient,
+):
+
+    init(
+        supertokens_config=SupertokensConfig("http://localhost:3567"),
+        app_info=InputAppInfo(
+            app_name="SuperTokens Demo",
+            api_domain="http://api.supertokens.io",
+            website_domain="http://supertokens.io",
+            api_base_path="/auth",
+        ),
+        framework="fastapi",
+        recipe_list=[
+            emailpassword.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
+        ],
+    )
+    start_st()
+
+    response_1 = sign_up_request(
+        driver_config_client, "random@gmail.com", "validpass123"
+    )
+    assert response_1.status_code == 200
+    dict_response = json.loads(response_1.text)
+    user_info = dict_response["user"]
+
+    link = await send_reset_password_email("public", "invalidUserId")
+    assert isinstance(link, SendResetPasswordEmailUnknownUserIdError)
+
+    with raises(GeneralError) as err:
+        await send_reset_password_email("invalidTenantId", user_info["id"])
+    assert "status code: 400" in str(err.value)
