--- tests/thirdpartypasswordless/__init__.py
@@ -1,3 +1,0 @@
-import nest_asyncio  # type: ignore
-
-nest_asyncio.apply()  # type: ignore
